Added Routing-url option: Forward all messages to that routing url
